Added a fairly no frills sort of progress bar widget.
Example usage.
set_widgets(self, {type:"progress", on_completion:"debug('progress finished!')", rect:[0,0,200,50], min:0, max:100, resolution:"double", frame_set:"default_button", id:"p1", color:"grey"})
map(range(5), schedule(value*50+50, add(self.widgets.p1.progress, 20)))

+++ src/poly_line_widget.hpp
@@ -85,7 +85,7 @@ protected:
 			}
 			calc_coords();
 		} else if(key == "width") {
-			width_ = v["width"].as_decimal().as_float();
+			width_ = GLfloat(v["width"].as_decimal().as_float());
 		} else if(key == "color") {
 			color_ = graphics::color(v).as_sdl_color();
 		}
